requirements.txt
asgiref==3.8.1
async-timeout==5.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==43.0.1
daphne==4.0.0
dj-database-url==2.3.0
Django==5.1.2
django-environ==0.11.2
django-filter==24.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
packaging==24.1
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dotenv==1.0.1
redis==5.1.1
service-identity==24.1.0
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
zope.interface==7.1.0
